refactor to cleanup types and follow are-report state config
