+++ apps/fishing-map/features/area-report/activity/ReportActivity.tsx
@@ -10,12 +10,12 @@ import {
   useReportFilteredTimeSeries,
 } from 'features/area-report/reports-timeseries.hooks'
 import { selectTimeComparisonValues } from 'features/area-report/reports.selectors'
-import { ReportActivityGraph } from 'types'
-import { selectReportActivityGraph } from 'features/app/selectors/app.reports.selector'
 import ReportActivityPlaceholder from 'features/area-report/placeholders/ReportActivityPlaceholder'
 import ReportActivityPeriodComparison from 'features/area-report/activity/ReportActivityPeriodComparison'
 import ReportActivityPeriodComparisonGraph from 'features/area-report/activity/ReportActivityPeriodComparisonGraph'
 import UserGuideLink from 'features/help/UserGuideLink'
+import { selectReportActivityGraph } from '../reports.config.selectors'
+import { ReportActivityGraph } from '../reports.types'
 import ReportActivityEvolution from './ReportActivityEvolution'
 import ReportActivityBeforeAfter from './ReportActivityBeforeAfter'
 import ReportActivityBeforeAfterGraph from './ReportActivityBeforeAfterGraph'
