+++ apps/fishing-map/features/area-report/activity/ReportActivityBeforeAfter.tsx
@@ -3,13 +3,13 @@ import { useTranslation } from 'react-i18next'
 import { useSelector } from 'react-redux'
 import { InputDate, InputText, Select, SelectOption } from '@globalfishingwatch/ui-components'
 import { useReportTimeCompareConnect } from 'features/area-report/reports-timecomparison.hooks'
-import { selectReportTimeComparison } from 'features/app/selectors/app.reports.selector'
 import { selectActiveActivityAndDetectionsDataviews } from 'features/dataviews/selectors/dataviews.selectors'
 import { getSourcesSelectedInDataview } from 'features/workspace/activity/activity.utils'
 import { selectReportAreaIds } from 'features/area-report/reports.selectors'
 import { selectDatasetAreaDetail } from 'features/areas/areas.slice'
 import { TrackCategory, trackEvent } from 'features/app/analytics.hooks'
 import { MAX_MONTHS_TO_COMPARE, MAX_DAYS_TO_COMPARE } from 'features/area-report/reports.config'
+import { selectReportTimeComparison } from '../reports.config.selectors'
 import styles from './ReportActivityBeforeAfter.module.css'
 
 export default function ReportActivityBeforeAfter() {
