+++ apps/fishing-map/features/area-report/activity/ReportActivityGraphSelector.tsx
@@ -8,14 +8,12 @@ import {
   REPORT_ACTIVITY_GRAPH_BEFORE_AFTER,
   REPORT_ACTIVITY_GRAPH_PERIOD_COMPARISON,
 } from 'data/config'
-import {
-  selectActiveReportDataviews,
-  selectReportActivityGraph,
-} from 'features/app/selectors/app.reports.selector'
+import { selectActiveReportDataviews } from 'features/app/selectors/app.reports.selector'
 import { useFitAreaInViewport } from 'features/area-report/reports.hooks'
-import { ReportActivityGraph } from 'types'
 import { useSetReportTimeComparison } from 'features/area-report/reports-timecomparison.hooks'
 import { TrackCategory, trackEvent } from 'features/app/analytics.hooks'
+import { selectReportActivityGraph } from '../reports.config.selectors'
+import { ReportActivityGraph } from '../reports.types'
 
 type ReportActivityGraphSelectorProps = {
   loading: boolean
