+++ apps/fishing-map/features/area-report/activity/ReportActivityPeriodComparison.tsx
@@ -3,7 +3,6 @@ import { useTranslation } from 'react-i18next'
 import { useSelector } from 'react-redux'
 import { InputDate, InputText, Select } from '@globalfishingwatch/ui-components'
 import { useReportTimeCompareConnect } from 'features/area-report/reports-timecomparison.hooks'
-import { selectReportTimeComparison } from 'features/app/selectors/app.reports.selector'
 import { selectActiveActivityAndDetectionsDataviews } from 'features/dataviews/selectors/dataviews.selectors'
 import { getSourcesSelectedInDataview } from 'features/workspace/activity/activity.utils'
 import { selectReportAreaIds } from 'features/area-report/reports.selectors'
@@ -12,6 +11,7 @@ import Hint from 'features/help/Hint'
 import { COLOR_PRIMARY_BLUE } from 'features/app/app.config'
 import { TrackCategory, trackEvent } from 'features/app/analytics.hooks'
 import { MAX_MONTHS_TO_COMPARE, MAX_DAYS_TO_COMPARE } from 'features/area-report/reports.config'
+import { selectReportTimeComparison } from '../reports.config.selectors'
 import styles from './ReportActivityBeforeAfter.module.css'
 
 export default function ReportActivityGraph() {
