+++ apps/fishing-map/features/vessel/vessel.config.selectors.ts
@@ -1,8 +1,8 @@
 import { createSelector } from '@reduxjs/toolkit'
 import { VesselIdentitySourceEnum } from '@globalfishingwatch/api-types'
-import { VesselProfileState, VesselProfileStateProperty } from 'types'
 import { selectQueryParam } from 'routes/routes.selectors'
 import { DEFAULT_VESSEL_STATE } from 'features/vessel/vessel.config'
+import { VesselProfileStateProperty, VesselProfileState } from './vessel.types'
 
 type VesselProfileProperty<P extends VesselProfileStateProperty> = Required<VesselProfileState>[P]
 export function selectVesselProfileStateProperty<P extends VesselProfileStateProperty>(
