+++ apps/fishing-map/types/index.ts
@@ -1,12 +1,5 @@
 import { BaseUrlWorkspace, UrlDataviewInstance } from '@globalfishingwatch/dataviews-client'
-import {
-  DatasetSubCategory,
-  DataviewCategory,
-  EventType,
-  GearType,
-  VesselIdentitySourceEnum,
-  VesselType,
-} from '@globalfishingwatch/api-types'
+import { EventType } from '@globalfishingwatch/api-types'
 import {
   DrawFeatureType,
   FourwingsVisualizationMode,
@@ -14,16 +7,14 @@ import {
   HEATMAP_LOW_RES_ID,
   RulerData,
 } from '@globalfishingwatch/deck-layers'
-import {
-  REPORT_VESSELS_GRAPH_GEARTYPE,
-  REPORT_VESSELS_GRAPH_FLAG,
-  REPORT_ACTIVITY_GRAPH_EVOLUTION,
-  REPORT_ACTIVITY_GRAPH_BEFORE_AFTER,
-  REPORT_ACTIVITY_GRAPH_PERIOD_COMPARISON,
-  REPORT_VESSELS_GRAPH_VESSELTYPE,
-} from 'data/config'
-import { SearchType } from 'features/search/search.config'
 import { MapAnnotation } from 'features/map/overlays/annotations/annotations.types'
+import { AreaReportState, AreaReportStateProperty } from 'features/area-report/reports.types'
+import { VesselProfileState, VesselProfileStateProperty } from 'features/vessel/vessel.types'
+import {
+  VesselGroupReportState,
+  VesselGroupReportStateProperty,
+} from 'features/vessel-groups/vessel-groups.types'
+import { VesselSearchState, VesselSearchStateProperty } from 'features/search/search.types'
 export { Locale } from '@globalfishingwatch/api-types'
 
 type WorkspaceViewportParam = 'latitude' | 'longitude' | 'zoom'
@@ -31,28 +22,16 @@ type WorkspaceTimeRangeParam = 'start' | 'end'
 export type BufferUnit = 'nauticalmiles' | 'kilometers'
 export type BufferOperation = 'dissolve' | 'difference'
 
-type ReportStateProperty =
-  | 'reportActivityGraph'
-  | 'reportAreaBounds'
-  | 'reportCategory'
-  | 'reportResultsPerPage'
-  | 'reportTimeComparison'
-  | 'reportVesselFilter'
-  | 'reportVesselGraph'
-  | 'reportVesselPage'
-  | 'reportBufferValue'
-  | 'reportBufferUnit'
-  | 'reportBufferOperation'
-
 export type WorkspaceStateProperty = keyof WorkspaceState
 type AppStateProperty = keyof AppState
 
-type AnyStateProperty = WorkspaceStateProperty | ReportStateProperty | AppStateProperty
+type AnyStateProperty = WorkspaceStateProperty | AppStateProperty
 
 export type WorkspaceParam =
   | WorkspaceViewportParam
   | WorkspaceTimeRangeParam
   | AnyStateProperty
+  | AreaReportStateProperty
   | VesselProfileStateProperty
   | VesselGroupReportStateProperty
   | VesselSearchStateProperty
@@ -61,29 +40,6 @@ export type WorkspaceViewport = Record<WorkspaceViewportParam, number>
 type WorkspaceTimeRange = Record<WorkspaceTimeRangeParam, string>
 
 type BivariateDataviews = [string, string]
-export type ReportActivityGraph =
-  | typeof REPORT_ACTIVITY_GRAPH_EVOLUTION
-  | typeof REPORT_ACTIVITY_GRAPH_BEFORE_AFTER
-  | typeof REPORT_ACTIVITY_GRAPH_PERIOD_COMPARISON
-
-export type ReportActivityTimeComparison = {
-  start: string
-  compareStart: string
-  duration: number
-  durationType: 'days' | 'months'
-}
-
-export enum ReportCategory {
-  Fishing = DatasetSubCategory.Fishing,
-  Presence = DatasetSubCategory.Presence,
-  Detections = DataviewCategory.Detections,
-  Environment = DataviewCategory.Environment,
-}
-
-export type ReportVesselGraph =
-  | typeof REPORT_VESSELS_GRAPH_GEARTYPE
-  | typeof REPORT_VESSELS_GRAPH_VESSELTYPE
-  | typeof REPORT_VESSELS_GRAPH_FLAG
 
 export interface WorkspaceState extends BaseUrlWorkspace {
   activityVisualizationMode?: FourwingsVisualizationMode
@@ -97,17 +53,6 @@ export interface WorkspaceState extends BaseUrlWorkspace {
   mapRulersVisible?: boolean
   daysFromLatest?: number // use latest day as endAt minus the number of days set here
   readOnly?: boolean
-  reportActivityGraph?: ReportActivityGraph
-  reportAreaBounds?: Bbox
-  reportCategory?: ReportCategory
-  reportTimeComparison?: ReportActivityTimeComparison
-  reportVesselFilter?: string
-  reportVesselGraph?: ReportVesselGraph
-  reportVesselPage?: number
-  reportBufferValue?: number
-  reportBufferUnit?: BufferUnit
-  reportBufferOperation?: BufferOperation
-  reportResultsPerPage?: number
   sidebarOpen?: boolean
   timebarGraph?: TimebarGraphs
   timebarSelectedEnvId?: string
@@ -115,68 +60,9 @@ export interface WorkspaceState extends BaseUrlWorkspace {
   visibleEvents?: VisibleEvents
 }
 
-export type VesselSearchState = {
-  query?: string
-  shipname?: string
-  sources?: string[]
-  searchOption?: SearchType
-  infoSource?: VesselIdentitySourceEnum
-  ssvid?: string
-  imo?: string
-  callsign?: string
-  codMarinha?: string
-  nationalId?: string
-  flag?: string[]
-  geartypes?: GearType[]
-  shiptypes?: VesselType[]
-  targetSpecies?: string
-  transmissionDateFrom?: string
-  transmissionDateTo?: string
-  owner?: string
-  fleet?: string[]
-  origin?: string
-}
-
-export type VesselSearchStateProperty = keyof VesselSearchState
-export type VesselSection = 'activity' | 'related_vessels' | 'areas' | 'insights'
-export type VesselAreaSubsection = 'fao' | 'eez' | 'mpa' | 'rfmo'
-export type VesselRelatedSubsection = 'encounters' | 'owners'
-export type VesselProfileActivityMode = 'voyage' | 'type'
-export type VesselProfileState = {
-  vesselDatasetId: string
-  vesselRegistryId?: string
-  vesselSelfReportedId?: string
-  vesselSection: VesselSection
-  vesselArea: VesselAreaSubsection
-  vesselRelated: VesselRelatedSubsection
-  vesselIdentitySource: VesselIdentitySourceEnum
-  vesselActivityMode: VesselProfileActivityMode
-  viewOnlyVessel: boolean
-}
-
-export type VesselProfileStateProperty = keyof VesselProfileState
-
-export type VesselGroupReportSection = 'vessels' | 'insights' | 'activity' | 'events'
-export type VesselGroupReportVesselsSubsection = 'flag' | 'shiptypes' | 'geartypes' | 'source'
-export type VesselGroupReportActivitySubsection = 'fishing-effort' | 'presence'
-export type VesselGroupReportEventsSubsection = 'fishing' | 'encounters' | 'port' | 'loitering'
-export type VesselGroupReportVesselsOrderProperty = 'shipname' | 'flag' | 'shiptype'
-export type VesselGroupReportVesselsOrderDirection = 'asc' | 'desc'
-
-export type VesselGroupReportState = {
-  viewOnlyVesselGroup: boolean
-  vesselGroupReportSection: VesselGroupReportSection
-  vesselGroupReportVesselsSubsection?: VesselGroupReportVesselsSubsection
-  vesselGroupReportActivitySubsection?: VesselGroupReportActivitySubsection
-  vesselGroupReportEventsSubsection?: VesselGroupReportEventsSubsection
-  vesselGroupReportVesselPage?: number
-  vesselGroupReportResultsPerPage?: number
-  vesselGroupReportVesselFilter?: string
-  vesselGroupReportVesselsOrderProperty?: VesselGroupReportVesselsOrderProperty
-  vesselGroupReportVesselsOrderDirection?: VesselGroupReportVesselsOrderDirection
-}
-
-export type VesselGroupReportStateProperty = keyof VesselGroupReportState
+export type AnyWorkspaceState = Partial<
+  WorkspaceState & AreaReportState & VesselProfileState & VesselGroupReportState
+>
 
 type RedirectParam = {
   'access-token'?: string
@@ -197,9 +83,10 @@ export type AppState = {
 }
 
 export type QueryParams = Partial<WorkspaceViewport> &
-  Partial<WorkspaceTimeRange> &
   WorkspaceState &
+  Partial<WorkspaceTimeRange> &
   Partial<VesselProfileState> &
+  Partial<AreaReportState> &
   Partial<VesselGroupReportState> &
   AppState &
   RedirectParam &
